feat(git): refactor git module structure and add error handling
