+++ src-tauri/src/lib.rs
@@ -136,14 +136,14 @@ pub fn run() {
             fs::search::fs_search,
             fs::grep::fs_grep,
             fs::grep::fs_glob,
-            git::git_resolve_repo,
-            git::git_status,
-            git::git_diff,
-            git::git_diff_content,
-            git::git_stage,
-            git::git_unstage,
-            git::git_commit,
-            git::git_push,
+            git::commands::git_resolve_repo,
+            git::commands::git_status,
+            git::commands::git_diff,
+            git::commands::git_diff_content,
+            git::commands::git_stage,
+            git::commands::git_unstage,
+            git::commands::git_commit,
+            git::commands::git_push,
             shell::shell_run_command,
             shell::shell_session_open,
             shell::shell_session_run,
