+++ src-tauri/src/modules/git/mod.rs
@@ -1,0 +1,7 @@
+pub mod commands;
+mod operations;
+mod parser;
+mod process;
+pub mod errors;
+pub mod types;
+pub mod utils;
