+++ src-tauri/src/modules/git/types.rs
@@ -1,0 +1,93 @@
+use std::time::Duration;
+use serde::Serialize;
+
+pub(crate) const DEFAULT_TIMEOUT_SECS: u64 = 30;
+pub(crate) const MAX_TIMEOUT_SECS: u64 = 120;
+pub(crate) const MAX_OUTPUT_BYTES: usize = 512 * 1024;
+pub(crate) const POLL_INTERVAL: Duration = Duration::from_millis(50);
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitRepoInfo {
+    pub repo_root: String,
+    pub branch: String,
+    pub upstream: Option<String>,
+    pub is_detached: bool,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitChangedFile {
+    pub path: String,
+    pub original_path: Option<String>,
+    pub index_status: String,
+    pub worktree_status: String,
+    pub staged: bool,
+    pub unstaged: bool,
+    pub untracked: bool,
+    pub status_label: String,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitStatusSnapshot {
+    pub repo_root: String,
+    pub branch: String,
+    pub upstream: Option<String>,
+    pub ahead: u32,
+    pub behind: u32,
+    pub is_detached: bool,
+    pub changed_files: Vec<GitChangedFile>,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitDiffResult {
+    pub diff_text: String,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitDiffContentResult {
+    pub original_content: String,
+    pub modified_content: String,
+    pub is_binary: bool,
+    pub fallback_patch: String,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitCommitResult {
+    pub commit_sha: String,
+    pub summary: String,
+}
+
+#[derive(Serialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GitPushResult {
+    pub remote: Option<String>,
+    pub branch: Option<String>,
+    pub pushed: bool,
+}
+
+pub(crate) struct GitOutput {
+    pub(crate) stdout: Vec<u8>,
+    pub(crate) stderr: Vec<u8>,
+    pub(crate) exit_code: Option<i32>,
+    pub(crate) timed_out: bool,
+}
+
+pub(crate) enum TextSource {
+    Missing,
+    Binary,
+    Text(String),
+}
+
+impl TextSource {
+    pub(crate) fn into_text(self) -> String {
+        match self {
+            TextSource::Text(text) => text,
+            TextSource::Missing | TextSource::Binary => String::new(),
+        }
+    }
+}
